% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{default_cache_location}
\alias{default_cache_location}
\title{Default OSF cache location}
\usage{
default_cache_location()
}
\value{
Path to default cache location
}
\description{
Evaluates to "_osfcache" in your R project root (determined by
\link[here:here]{here::here}).
}
