% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-utils.R
\name{osf_guid}
\alias{osf_guid}
\title{Get the GUID of an OSF node}
\usage{
osf_guid(osf_tbl)
}
\arguments{
\item{osf_tbl}{An \code{osf_tbl} metadata object that is associated with an object
in the OSF cache.}
}
\value{
The object's GUID
}
\description{
Get the GUID of an OSF node
}
