% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-utils.R
\name{osf_last_updated}
\alias{osf_last_updated}
\title{Get when an OSF node was last updated}
\usage{
osf_last_updated(osf_tbl)
}
\arguments{
\item{osf_tbl}{An \code{osf_tbl} metadata object that is associated with an object
in the OSF cache.}
}
\value{
The date when last modified
}
\description{
Get when an OSF node was last updated
}
