% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-utils.R
\name{osf_path}
\alias{osf_path}
\title{Get the location of a cached OSF download}
\usage{
osf_path(osf_tbl)
}
\arguments{
\item{osf_tbl}{An \code{osf_tbl} metadata object that is associated with an object
in the OSF cache.}
}
\value{
The object's local path
}
\description{
Get the location of a cached OSF download
}
