% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{osfcache_get}
\alias{osfcache_get}
\title{Retrieve an object from the OSF cache}
\usage{
osfcache_get(
  guid,
  cache_dir = default_cache_location(),
  create_cache_dir = TRUE,
  ...
)
}
\arguments{
\item{guid}{The GUID on OSF for the desired object.}

\item{cache_dir}{The OSF cache location.}

\item{create_cache_dir}{Create the cache directories if they don't exist.}

\item{...}{Any parameters passed to \code{osfr::osf_download}}
}
\value{
The current \code{osf_tbl} associated with the requested object
}
\description{
By default, objects downloaded from OSF will be stored in the "osfcache"
folder. In that folder there are two folders: "meta" and "assets".
The "meta" stores serialized \code{osfr} metadata objects, and the
"assets" folder keeps the downloads themselves. If the requested object
doesn't exist in the cache or is outdated, the most recently updated
object will be downloaded.
}
