% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{osfcache_is_outdated}
\alias{osfcache_is_outdated}
\title{Check to see if cache is outdated}
\usage{
osfcache_is_outdated(
  guid,
  cache_dir = default_cache_location(),
  create_cache_dir = TRUE
)
}
\arguments{
\item{guid}{An OSF GUID string}

\item{cache_dir}{Where the cache is located}

\item{create_cache_dir}{If cache dir doesn't exist it, create it if TRUE}
}
\value{
Logical if the cache is outdated
}
\description{
Fetches the most recent meta with GUID \code{guid} to see if
\enumerate{
\item the requested object exists in the cache
\item the cache's version of the object is outdated
}
}
