% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{password_decrypt}
\alias{password_decrypt}
\title{Encrypt an R object with a password}
\usage{
password_decrypt(x, password)
}
\arguments{
\item{x}{Either
\itemize{
\item A path to an Rds file for an encrypted ciphertext
\item An encrypted ciphertext object
}}

\item{password}{A character string}
}
\value{
An decrypted R object from the cipertext
}
\description{
This routine encrypts any R object with a password using
\href{https://cran.r-project.org/web/packages/sodium/vignettes/intro.html#secret-key-encryption}{sodium's secret key encryption}.
}
