% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{password_encrypt}
\alias{password_encrypt}
\title{Encrypt an R object with a password}
\usage{
password_encrypt(obj, password, file = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{Any object}

\item{password}{A character string}

\item{file}{If not \code{NULL}, a path to where the Rds object should be saved}

\item{verbose}{Signals to message() when/where an Rds file is written}
}
\value{
The cipertext bytes of the encrypted object
}
\description{
This routine encrypts any R object with a password using
\href{https://cran.r-project.org/web/packages/sodium/vignettes/intro.html#secret-key-encryption}{sodium's secret key encryption}.
}
